+++ src/app/features/catalogos/pais/pais.component.ts
@@ -4,11 +4,14 @@ import {PaisService} from '../../../core/services/pais.service';
 import {MatDialog} from '@angular/material/dialog';
 import {Validators} from '@angular/forms';
 import {Field, FormDialogGenericoComponent} from '../../../shared/form-dialog-generico/form-dialog-generico.component';
-import {MatButton} from '@angular/material/button';
+import {MatButton, MatIconButton} from '@angular/material/button';
 import {MatCard, MatCardContent, MatCardHeader} from '@angular/material/card';
 import {MatIcon} from '@angular/material/icon';
 import {MatToolbar} from '@angular/material/toolbar';
 import {NgIf} from '@angular/common';
+import {MatDrawer, MatDrawerContainer} from '@angular/material/sidenav';
+import {FormGenericoComponent} from '../../../shared/form-generico/form-generico.component';
+import {GenericoService} from '../../../core/services/generico.service';
 
 export interface ColumnasTabla {
   clave: string,
@@ -28,38 +31,42 @@ export interface ColumnasTabla {
     MatIcon,
     MatToolbar,
     NgIf,
+    MatDrawerContainer,
+    FormGenericoComponent,
+    MatDrawer,
+    MatIconButton,
   ],
   templateUrl: './pais.component.html',
   styleUrl: './pais.component.css',
 })
 export class PaisComponent implements OnInit {
   paisList: any[] = [];
-
-  columns : ColumnasTabla[] = [
-    {clave: 'fechaRegistro', valor: 'Fecha de registro', tipo: "fecha"},
-    {clave: 'clave', valor: 'Clave', tipo: "texto"},
-    {clave: 'nombre', valor: 'Nombre', tipo: "texto"},
-    {clave: 'descripcion', valor: 'Descripción', tipo: "texto"}
-  ];
-
-  actions = [
-    {name: 'Editar', icon: "edit", callback: (item: any) => this.openFormDialog(item)},
-    {name: 'Eliminar', icon: "delete", callback: (item: any) => this.delete(item)},
-    {name: 'Ver detalle', icon: "visibility", callback: (item: any) => this.verDetalle(item)}
-  ];
-
+  totalRecords: number = 0;
+  fieldsFilters!: Field[];
   queryParams = {
     max: 10,
     offset: 0,
     filtroGeneral: '',
     registrosEliminados: false,
   };
-  totalRecords: number = 0;
+  columns : ColumnasTabla[] = [
+    {clave: 'fechaRegistro', valor: 'Fecha de registro', tipo: "fecha"},
+    {clave: 'clave', valor: 'Clave', tipo: "texto"},
+    {clave: 'nombre', valor: 'Nombre', tipo: "texto"},
+    {clave: 'descripcion', valor: 'Descripción', tipo: "texto"},
+    {clave: 'activo', valor: 'Estatus', tipo: "boleano"},
+  ];
+  actions = [
+    {name: 'Editar', icon: "edit", tooltipText: 'Editar', callback: (item: any) => this.openFormDialog(item)},
+    {name: 'Eliminar', icon: "delete", tooltipText: 'Eliminar', callback: (item: any) => this.delete(item)},
+    {name: 'Recuperar eliminado', icon: "restore_from_trash", tooltipText: 'Recuperar registro eliminado', callback: (item: any) => this.recoverRegister(item.id)}
+  ];
 
   @ViewChild('tablaGenerica') tablaGenerica!: TablaGenericaComponent;
 
   constructor(
     private paisService: PaisService,
+    private genericoService: GenericoService,
     private dialog: MatDialog
   ) {
   }
@@ -80,21 +87,49 @@ export class PaisComponent implements OnInit {
       .subscribe((paises: any) => {
         this.paisList = paises.data;
         this.totalRecords = paises.count;
-        this.paisList = this.datosTabla;
       });
   }
 
+  formFiltros(): void {
+    this.fieldsFilters = [
+      {
+        name: 'filtroGeneral',
+        label: 'Clave / Nombre',
+        type: 'text',
+      },
+      {
+        name: 'registrosEliminados',
+        label: 'Ver eliminados',
+        type: 'toggle',
+      },
+    ]
+  }
+
+  procesarfiltros(form: any) {
+    this.queryParams= ({...this.queryParams, ...form});
+    this.lista();
+  }
+
+  resetFormFiltros(){
+    this.queryParams = {
+      max: 10,
+      offset: 0,
+      filtroGeneral: '',
+      registrosEliminados: false,
+    };
+    this.lista()
+  }
 
   // Transformar los datos para la tabla
-  get datosTabla(): any {
-    return this.paisList.map(pais => ({
-      id: pais?.id,
-      fechaRegistro: pais?.fechaRegistro,
-      clave: pais?.clave,
-      nombre: pais?.nombre,
-      descripcion: pais?.descripcion,
-    }));
-  }
+  // get datosTabla(): any {
+  //   return this.paisList.map(pais => ({
+  //     id: pais?.id,
+  //     fechaRegistro: pais?.fechaRegistro,
+  //     clave: pais?.clave,
+  //     nombre: pais?.nombre,
+  //     descripcion: pais?.descripcion,
+  //   }));
+  // }
 
   public handlePageChange(event: any) {
     this.queryParams.max = event.max;
@@ -138,13 +173,28 @@ export class PaisComponent implements OnInit {
     });
   }
 
-  delete(objeto: any) {
+  private async delete(objeto: any) {
+    const isConfirmed = await this.genericoService.confirmDialog(
+      '¿Está seguro de eliminar el registro?'
+    );
+    if (!isConfirmed) {
+      return;
+    }
+
     this.paisService.delete(objeto.id).subscribe(() => {
       this.lista();
     });
   }
 
-  verDetalle(item: any) {
-    alert(`ver Detalle ${item.nombre}`)
+  private async recoverRegister(id: number) {
+    const isConfirmed = await this.genericoService.confirmDialog(
+      '¿Está seguro de recuperar el registro?'
+    );
+    if (!isConfirmed) {
+      return;
+    }
+    this.paisService.recover(id).subscribe(() => {
+      this.lista(true);
+    });
   }
 }
