+++ src/app/features/catalogos/regimen-fiscal/regimen-fiscal.component.ts
@@ -6,11 +6,14 @@ import {Validators} from '@angular/forms';
 import {Field, FormDialogGenericoComponent} from '../../../shared/form-dialog-generico/form-dialog-generico.component';
 import {RegimenFiscalService} from '../../../core/services/regimenFiscal.service';
 import {ColumnasTabla} from '../pais/pais.component';
-import {MatButton} from '@angular/material/button';
+import {MatButton, MatIconButton} from '@angular/material/button';
 import {MatCard, MatCardContent, MatCardHeader} from '@angular/material/card';
 import {MatIcon} from '@angular/material/icon';
 import {MatToolbar} from '@angular/material/toolbar';
 import {NgIf} from '@angular/common';
+import {GenericoService} from '../../../core/services/generico.service';
+import {MatDrawer, MatDrawerContainer} from '@angular/material/sidenav';
+import {FormGenericoComponent} from '../../../shared/form-generico/form-generico.component';
 
 @Component({
   selector: 'app-regimen-fiscal',
@@ -23,13 +26,25 @@ import {NgIf} from '@angular/common';
     MatCardHeader,
     MatIcon,
     MatToolbar,
-    NgIf
+    NgIf,
+    MatDrawerContainer,
+    FormGenericoComponent,
+    MatDrawer,
+    MatIconButton
   ],
   templateUrl: './regimen-fiscal.component.html',
   styleUrl: './regimen-fiscal.component.css'
 })
 export class RegimenFiscalComponent implements OnInit{
+  totalRecords: number = 0;
   regimenFiscalList: any[] = [];
+  fieldsFilters!: Field[];
+  queryParams = {
+    max: 10,
+    offset: 0,
+    filtroGeneral: '',
+    registrosEliminados: false,
+  };
   columns : ColumnasTabla[] = [
     {clave: 'fechaRegistro', valor: 'Fecha de registro', tipo: "fecha"},
     {clave: 'clave', valor: 'Clave', tipo: "texto"},
@@ -37,29 +52,19 @@ export class RegimenFiscalComponent implements OnInit{
     {clave: 'descripcion', valor: 'Descripción', tipo: "texto"},
     {clave: 'activo', valor: 'Activo', tipo: "boleano"}
   ];
-
   actions = [
-    {name: 'Editar', icon: "edit", callback: (item: any) => this.openFormDialog(item)},
-    {name: 'Eliminar', icon: "delete", callback: (item: any) => this.delete(item)},
-    {name: 'Ver detalle', icon: "visibility", callback: (item: any) => this.verDetalle(item)}
+    {name: 'Editar', icon: "edit", tooltipText: 'Editar', callback: (item: any) => this.openFormDialog(item)},
+    {name: 'Eliminar', icon: "delete", tooltipText: 'Eliminar', callback: (item: any) => this.delete(item)},
+    {name: 'Recuperar eliminado', icon: "restore_from_trash", tooltipText: 'Recuperar registro eliminado', callback: (item: any) => this.recoverRegister(item.id)}
   ];
-
-  queryParams = {
-    max: 10,
-    offset: 0,
-    filtroGeneral: '',
-    registrosEliminados: false,
-  };
-
-  totalRecords: number = 0;
 
   @ViewChild('tablaGenerica') tablaGenerica!: TablaGenericaComponent;
 
   constructor(
     private regimenFiscalService: RegimenFiscalService,
+    private genericoService: GenericoService,
     private dialog: MatDialog
-  ) {
-  }
+  ) {}
 
   ngOnInit() {
     this.lista();
@@ -78,6 +83,36 @@ export class RegimenFiscalComponent implements OnInit{
         this.regimenFiscalList = regimenFiscalList.data;
         this.totalRecords = regimenFiscalList.count;
       });
+  }
+
+  formFiltros(): void {
+    this.fieldsFilters = [
+      {
+        name: 'filtroGeneral',
+        label: 'Clave / Nombre',
+        type: 'text',
+      },
+      {
+        name: 'registrosEliminados',
+        label: 'Ver eliminados',
+        type: 'toggle',
+      },
+    ]
+  }
+
+  procesarfiltros(form: any) {
+    this.queryParams= ({...this.queryParams, ...form});
+    this.lista();
+  }
+
+  resetFormFiltros(){
+    this.queryParams = {
+      max: 10,
+      offset: 0,
+      filtroGeneral: '',
+      registrosEliminados: false,
+    };
+    this.lista()
   }
 
   public handlePageChange(event: any) {
@@ -126,20 +161,41 @@ export class RegimenFiscalComponent implements OnInit{
     dialogRef.componentInstance.submitForm.subscribe(result => {
       if (data.id) {
         result = ({...result, id: data.id})
-        this.regimenFiscalService.update(result).subscribe(() => this.lista());
+        this.regimenFiscalService.update(result).subscribe((respueta) => {
+          if(respueta) this.genericoService.openSnackBar('Registro actualizado exitosamente', 'Aceptar', 'snack-bar-success', () =>{});
+          this.lista();
+        });
       } else {
-        this.regimenFiscalService.create(result).subscribe(() => this.lista());
+        this.regimenFiscalService.create(result).subscribe((respueta) => {
+          if(respueta) this.genericoService.openSnackBar('Registro creado exitosamente', 'Aceptar', 'snack-bar-success', () =>{});
+          this.lista();
+        });
       }
       dialogRef.close();
     });
   }
 
-  delete(objeto: any) {
+  private async delete(objeto: any) {
+    const isConfirmed = await this.genericoService.confirmDialog(
+      '¿Está seguro de eliminar el registro?'
+    );
+    if (!isConfirmed) {
+      return;
+    }
     this.regimenFiscalService.delete(objeto.id).subscribe(() => {
       this.lista();
-    });  }
+    });
+  }
 
-  verDetalle(item: any) {
-    alert(`ver Detalle ${item.nombre}`)
+  private async recoverRegister(id: number) {
+    const isConfirmed = await this.genericoService.confirmDialog(
+      '¿Está seguro de recuperar el registro?'
+    );
+    if (!isConfirmed) {
+      return;
+    }
+    this.regimenFiscalService.recover(id).subscribe(() => {
+      this.lista(true);
+    });
   }
 }
